Feat: add login guard and user check (#11)

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,12 @@ use App\Http\Requests\PostRequest; // 追記
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')
+             ->only(['store', 'edit', 'update', 'destroy']);
+    }
+
     public function store(PostRequest $request) {
         // $this->validate($request, [
         //     'body' => 'required|max:255'
@@ -32,10 +38,18 @@ class PostsController extends Controller
     }
 
     public function edit(Post $post) {
+        if ($post->user_id !== Auth::id()) {
+            return redirect(url('/posts', $post));
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
     public function update(PostRequest $request, Post $post) {
+        if ($post->user_id !== Auth::id()) {
+            return redirect(url('/posts', $post));
+        }
+
         $post->body = $request->body;
         if ($request->photo) {
             $post->deleteImage();
@@ -47,6 +61,10 @@ class PostsController extends Controller
     }
 
     public function destroy(Post $post) {
+        if ($post->user_id !== Auth::id()) {
+            return redirect(url('/posts', $post));
+        }
+
         $post->deleteImage();
         $post->delete();
         return redirect('/home');
